Remove `areChunksLoaded` check to avoid double get chunk
In `moonrise$areChunksLoaded`, it pre loop chunks in range and do fullchunk.get calls to check if the chunk is loaded. But get chunk calls are expensive, then I moved the check in the following loop to prevent double call it.

can be around 1 to 2 more faster in test (8192 stacked minecarts with 1min spark profiling).